src/a-just-for-test/requirement.txt
openai
fastapi
fastapi
crewai
crewai_tools
python-dotenv